django/api/requirements.txt
Django==3.1.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
djoser==2.1.0
django-cors-headers==3.7.0
social-auth-app-django==4.0.0
drf-yasg==1.20.0
requests==2.25.1
twilio==6.56.0
Pillow==8.2.0
django-environ==0.4.5
django-admin-tools==0.9.1
django-admin-charts==0.24.1
django-ckeditor==6.0.0